:truck: :recycle: :white_check_mark: get_contradicting_block_header.spec.ts to db.spec.ts
